Add new `Style/RedundantToS` cop
This new cop looks for instances where `to_s` is redundant -- where it's chained to a string (including interpolated, macro, heredoc and `String.new`), or to another `to_s` call.

Although it's unlikely to happen, the cop is marked as `SafeAutoCorrect: false` because `to_s` could be overridden on the receiver or `String` to not return a string.